DeskTop: Fix resource leak opening/closing windows
A list of FileRecords (32 byte structs) are is held for each open
window; a list is created when the window is opened, and freed (with
compaction) when the window is closed. The mapping of window to list
(and hence, space to reclaim) was done using the icon associated with
the window (i.e. volume/folder icon). For folders, this meant that if
the folder's parent window was closed first the icon->window mapping
for the folder would be lost and the space would leak, eventually
leading to the inability to open windows.

Change the tracking to use the window id instead. Fixes #19