DeskTop: Fix drag/drop onto window. Fixes #219
Path calculations were incorrect when dropping onto a window. This was
introduced in 1c882b3 when leading/trailing spaces were removed from
the icon names. The drop-on-window code path appended an extra '/';
previously, the drop-on-icon code paths appended the trailing ' ', and
a common code path removed it. The common code path was simplified,
leading to the extra '/' sticking around. Remove it in the
drop-on-window code path.