DeskTop: Allow arbitrary types in Apple menu. Resolves #293
DAs with high bit set in auxtype are still filtered out.

It works with SYS, S16, BAS and BIN files, as well as anything handled
by a BASIS.SYSTEM if present. This does not support preview types,
e.g. images, text, fonts - they will appear in the menu but nothing
happens when invoked.