DeskTop: Rework DA invocation to simplify selection access. Resolves #295
Preview DAs are now invoked with the full path at $220.

As a convenience, DAs invoked directly from the Apple menu are also
given the path to a file selection, if any, at $220 (set to 0 if
none). This simplifies the logic for Preview DAs and also supports
putting previewable file types in the Apple menu, while still allowing
preview DAs to manually be placed in DESK.ACC/ to support direct
invocation on other file types.