DeskTop & Selector: Fix sorting of files to accomodate GS/OS names. Fixes #384
The sorting logic for the file browser code (and View > by Name)
relied on a "sentinel" value of "ZZZZZZZZZZZZZZZ" which would sort
after all names. This isn't valid (since lowercase > uppercase).

Use a string with $7F instead, which isn't even a valid filename
character. (Yes, 15-letter long names are overkill here, but it's a
minimal change.)