DeskTop: Erase before redrawing header when opening via path. Fixes #392
Edge case for the fix in ac5c258157fa222e81cbc0acb1ec768dc89a30f7 -
if the garbage data drawn temporarily in the header is too long
(e.g. a large volume) the redraw wouldn't clear it. Erase the
header area before redrawing it.

(It's a horrible hack to begin with.)