DeskTop: Close Delete an Entry after deleting. Fixes #190
The original behavior for Selector > Delete an Entry was that on OK
the entry would be removed, but the dialog rould remain open. Cancel
would close the dialog, but not cancel any deletions. This was
convenient if you understood the behavior and wanted to delete
multiple entries, but confusing otherwise.

Now the dialog closes after deleting the item, as the UI would
indicate.

(An alternate fix would be to change the button labels to "Delete
Entry" and "Close", and change the menu item to "Delete Entries" or
something, but that's more intrusive.)