Use SmartPort STATUS calls to determine drive types. Fixes #437, #438, #439
Both DESKTOP.SYSTEM and DeskTop itself inspected firmware byte $CnFB
to determine the "type" of device. This failed for the IIgs /RAM5
volume since the firmware for Slot 5 is a general SmartPort device.

Replace this with actual SmartPort STATUS calls and inspect the Device
Information Block that is returned. This allow DeskTop to copy itself
to /RAM5 on startup, and show /RAM5 with a RAMDisk icon.

Note that the DIB's Device Type byte for other disk types (e.g. 3.5"
and 5.25" floppies) otherwise does not appear trustworthy, at least in
emulators, so the heuristic of 800k=3.5" is retained. This is needed
for generic block devices anyway.

Also, since it was easy while there, show 140k devices as 5.25"
floppies.