Selector: Don't crash if IIgs control panel invoked during alert. Fixes #443
The alert code runs with ALTZPON set, as it resides in Aux LCBANK1.
If the IIgs Control Panel is invoked during this state, the main stack
is corrupted therefore a crash occurs when the alert is closed.

DeskTop wasn't affected by this, and there's a magic incantation that
prevents this - setting some bytes in the Alt ZP. Unknown why this works
but it does.