DeskTop: Avoid off-by-one in "Copy to RAMCard..." dialog. Fixes #462
The "get size or count files" common logic counts the number of files in a selection. It includes a root folder, but not a root volume, since that's not a file. So far so good.

But the "Copy to RAMCard..." dialog used for an "on first use" Selector entry uses this to calculate the number of files to show. Seems okay, right? But if the Selector entry is in a volume dir (e.g. /Vol/Entry.Name), the parent (e.g. /Vol) is not counted but it is copied (as a subdir), so there's an off-by-one and the counter decrements past zero.

To fix, detect this case (copy target is a volume directory), and increment the counter.