DeskTop: Include actual icon height in scroll/bbox calculations. Fixes #414
Icons have variable height, but the scroll/bbox calculations used a max
height value. This caused the bbox to be *larger* when shorter icons
were present, since the top-left of icon is offset down.

Use the actual icon height in calculations, which makes the bbox more
accurate and prevents scrollbar from appearing unnecessarily in some
cases.