DeskTop: Fix remaining count in copy/move dialog for folders. Fixes #470
For copy/move operations, when processing a directory, the files
remaining count was decremented before the "show remaining count minus
one" routine was executed, so if a directory was the last thing copied
it could wrap to 65535. In other cases, the decrement was after the
routine.

Switch the order to match the other cases.