DeskTop: Prevent memory corruption dragging onto short folder names. Fixes #471
When dragging/dropping, the target needs to have its free/used numbers
updated. When the target is an icon, this is done by looking for
related windows and updating all of them.

The logic for finding "related windows" composed a path, assuming the
target was a volume, e.g. "/HD". But this logic was flawed - if the
target was not a volume, this is not a useful prefix (e.g. "/SUBDIR").
Worse, the code hadn't been updated to account for icon names no
longer having whitespace prefix/suffix, so the path was bogus. If the
folder name was one character (e.g. "A") then 256 bytes would be
copied into a 65-byte buffer, overwriting code.

Two fixes:

* The logic for composing the path from a volume icon name was fixed.
* Logic was added so that if the target was not a volume icon, the path
  prefix was computed correctly using the icon's window's path.