DeskTop: Set explicit portmap when drawing alert. Fixes #486
MGTK is non-intuitive. It would seem like you can draw to the screen
by calling InitPort (to initialize a grafport to the defaults) then
SetPort (to select the grafport). But SetPort actually applies the
current viewport to the selected port.

Operations therefore implicitly depend on having a non-clipped port
left behind unless they call SetPortBits. So... call that explicitly
from the alert code.