Selector: Don't crash/hang if SELECTOR.LIST is absent. Fixes #485
Tentative fix for #485, which is a report about Selector hanging
on load. I was able to repro if the SELECTOR.LIST file was missing,
but I'm not sure that's what the reporter was seeing.

Selector didn't error check when loading the file, so uninitialized
memory would be used for the run lists, leading to crashes/hangs.
Fix by initializing the run list header in case of failure.