Disk Copy: Fix alert buttons. Fixes #492
The switch to the common alert button lib in b748c7bb didn't account
for the need for a MGTK call relay. Adapting this was made more
complicated by the need for self-modifing call sites.

Fix the lib to use a LIB_MGTK_CALL, define that where needed, and
switch the Disk Copy auxlc's MGTK relay to grab the params from the
call site to support self-modifying code. As a side effect, since this
removes 3 bytes from each MGTK call site, this saves over 400 bytes!