DeskTop: Fix error after file copy
The fix in 29f0902d expanded a filename buffer, but there was an
overlapping data buffer just before it used during file copies that
was not adjusted. This resulted in an error at the end of a file
copy when file info was being applied to the destination path.

Shrink that buffer to avoid the collision.