Don't touch Le Chat Mauve Eve registers unless detected. Fixes #280
The Le Chat Mauve Eve RGB card and Uthernet II have overlapping
registers if installed in Slot 3. Unconditionally poking the
overlapping registers can make the Uthernet II unhappy.

Incorporate a detection routine based on the one from the LCM Eve
manual which detects the LCM Eve with non-overlapping registers, and
use this in both DeskTop and Selector.

This is a speculative fix since I don't have either an LCM Eve or an
Uthernet II.