Launcher: Fix error with copy-to-RAMCard logic. Fixes #509
The logic for copying to RAMCard relies on enumerating directories.
This reads file entries from the directory one at a time. Directory
data is structured as pseudo-blocks, and when the last entry in a
block, data needs to be skipped to advance to the next block.

If there is a next block, theres 1 byte of padding and 4 bytes for the
next/prev block pointers. But if there are no more blocks, only 1 more
byte can be read. The code was verifying that 5 bytes were read, and
failing if that were not the case. If the last file in the block had
just be read, then it would fail to be copied.

Fix this logic (rely on EOF to tell when we're done), and remove some
magic numbers.