DeskTop: Fix initial window widths. Fixes #514
The initial window size is determined by computing the bounding box of
the icons (in screen space), then clamping to min/max width and
height. This requires calculating the box height from bottom - top,
but the corresponding calculation for width from right - left was
missing, so subsequent windows were wider.