DeskTop: Fix highlighting after non-menu DA invocation. Fixes #516
Full screen DAs use JUMP_TABLE_HILITE_MENU to invert the menu item
used to invoke them. This is done so that screen dump shows no
highlight, and so screen savers display the menu with an invert that
gets re-inverted when the command finishes.

This extra inversion causes problems when the DA was not invoked using
the menu, e.g. using Apple+Open or double-click. To fix, clear the
last menu id when the menu invocation code has completed, and turn the
HILITE_MENU call into a no-op.