DeskTop: Reclaim space using MLI-style passing for calls.
Building on the previous commit, use MLI-style passing of arguments in
a few more cases. This is where the call number and parameter list
address follow the JSR and are accessed using the stack pointer,
eliminating 3 bytes per call site (for the LDA, LDX, LDY).

* For ProDOS MLI calls (need to bank in Main ZP / ROM)
* For IconTK (need to bank in Aux)

This saves ~500 bytes in Main.

In addition, within the aux/main relays in the LC area, use helper
procs for banking, which saves another ~80 bytes.