MGTK: Tentative fix for text clipping glitch. Fixes #505
In DrawTextImpl when width_bytes is incremented when left_mod14 +
clipped_left is negative. Later on, in the "For an on-screen
destination" block, width_bytes is doubled. The error case appears to
be that width_bytes ends up one greater than it should in this case
i.e. that initial increment is doubled.

Deferring the increment until after the doubling seems to resolve the
problem, but I don't claim I fully understand what's going on.