DeskTop: Improve edge cases at max icons. Fixes #528
DeskTop supports up to 126 icons, but a warning was shown at the 124
mark. Note that the limit in the code is 127 but due to off-by-one the
127th icon can't be shown. This can probably be fixed, but for now the
docs are updated to say 126.

There are also two warning message identifiers, both with the same
message. One is shown in the case where no windows are open (i.e.
the directory has too many icons), but both say windows need to
be closed. Oops. I left a TODO for changing the message.

The other one had a useless Cancel button, per #528. Removed.