DeskTop: Handle max icon case for File > New Folder. Fixes #527
If the maximum number of icons are already present, then when File >
New Folder attempts to refresh the active window it will fail. A
warning was shown, but the stack-restore exit from the fresh proc was
not handled, leading to memory corruption.

Fix this by catching the error, and closing the window.