DeskTop: Enable up to 127 icons. Fixes #529
Fix some off-by-one calculations around the maximum number of icons.

Note that 127 is the max number of icons, but DeskTop will reserve
icons for all ProDOS devices, even if there isn't a volume on-line, so
if you have two drives, but one drive is empty, that's one icon for
the on-line volume, one reserved icon, and so only 125 file icons
could be shown.