DeskTop: Update Selector menu from file content. Fixes #526
If the SELECTOR.LIST file is modified out from under DeskTop
(not really supported, but sure) then DeskTop can get into an
inconsistent state because the menu and file resources aren't
in sync.

Since the file is loaded and modified correctly for the add, edit and
remove actions, just change the logic to always repopulate the menu
items after the file contet is modified.