DeskTop: Only require Apple key to run BIN when double-clicking.
DeskTop requires a modifier when launching a BIN file, presumably as a
safety measure since many BIN files are data not executable. But the
behavior was confusing - Apple-O would work (since a modifier was
down), and File>Open could use a modifier.

Simplify the logic and only require an Apple key when double-clicking.
Assume that for menu commands (including the keyboard shortcuts for
the Selector menu) are intentional.

Fixes #530, #531