DeskTop: Handle ProDOS error $2F when ejecting disks. Fixes #536
After ejecting a disk, ProDOS reports error $2F ("Device off-line") in
response to ON_LINE in some cases. This error code is only documented
in Tech Note 21, and the original A2D code didn't handle it. It should
be treated the same as error $27 ("I/O error"), and just cause the
attempt to refresh the icon to fail.