DeskTop: Fix max icon calculations. Fixes #537
A couple of issues:

(1) Calculation for "how many file records are free?" was off by one
if there were no file records, since the dir counts will always
consume some bytes too. Given that max # windows is < size of file
record, always one record is wasted. So max is 127, which equals
kMaxIconCount.

(2) DEVCNT was used to reserve icons for devices, but (a) DEVCNT is
one less than the number of devices and (a) it doesn't include Trash,
which takes an icon.

Fixing this will reduce the number of icons available for files,
but ensures the right number of icons are reserved for devices.