DeskTop: Show errors when two volumes have the same name. Fixes #542
The original A2D code tried to detect ProDOS error $57 from ON_LINE
during volume icon creation, but that is only delivered when files
are open. Add a manual check for duplicate volume names, and don't
create the second.

Also, if a volume is renamed, handle the error and show an alert
there as well.

Note that ProDOS is still going to be confused by the duplicate
volumes, since it can't really handle this situation.

(This turned up an off-by-one glitch in the rename code as well,
where the first renamed volume was processed twice!).