DeskTop: Replace large "Warning!" dialog with standard alert. Fixes #207
The benefits are (1) code size reduction in main and (2) the alerts
use the "save area" to save/restore the background, so occluded
windows don't need to repaint.

This affects the following warnings:

* "Please insert the system disk."
* "The list is full. You must delete an entry before you can add new entries."
* "A window must be closed before opening this item."
* "There are too many windows open on the desktop!"
* "Do you want to save the new list on the system disk?"

The alert mechanism (in DeskTop) now handles two-line messages and is
passed a "params" block (via A,X). This is not yet used except by
shims for the previous ID-based alert calls and ID-based warning
calls, which now populate this block.

Future work includes making the string line-wrap instead of requiring
two strings to be passed in, unifying with the Disk Copy and Selector
versions of the code, and deduping DeskTop's two ID-based paths.