Unify Alert code between Selector and DeskTop.
Split the Selector alert code into two parts: option parsing, and
showing the alert, with the first part passing a "params" block (via
A,X). With some additional refactoring, this means that the core
of the Selector and DeskTop alert code is now identical, allowing
it to be moved into a lib/ in the future.

Disk Copy will be slightly more complicated to unify, since it
has additional behaviors: Yes/No buttons, and auto-dismiss on
disk insertion.

No behavior changes.