Build: Set TERM to xterm-256color in makefile
If TERM is something boring like vt100 then setaf fails. Arbitrarily
setting TERM is not really correct, but practically speaking every
terminal users are likely to use and every build environment these
days is xterm-256color compatible. If anyone really complains we can
make the color coding calls used in Makefiles and utils smarter.

Now that it wouldn't be too much duplication in the build file, switch
the GitHub action to build each language as a separate step, so that
it's easier to see if e.g. a non-English build failed due to string
lengths.