DeskTop: Fix list views
This was broken in 70568001 when cached_window_icon_count was zeroed
out. It's not actually an icon count, but an entry count. (Rename
coming in a future rev.) Zeroing it out is correct when a window is
being refreshed or destroyed, but it was being zeroed out when the
view type was switched to a non-icon view, so the list view would be
empty.

Given that it was broken in alpha34 and no-one reported it (until
now), it shows how useless the list views are. :(