This Apple DA: Improve Mockingboard detection when accelerated.
Mockingboard detection is done with a timing loop, so an accelerator
can cause it to fail. James N. suggested a trick used elsewhere to
slow accelerators: tickle slot 6. Since slot 6 is the usual home for a
floppy controller, accelerators commonly treat it as a "slow slot" and
slow down for compatibility with timing-sensitive Disk II routines.

The detection improvement was confirmed with a real Zip Chip, as well
as MAME, via the awesome Ample.