DeskTop: Make selection when opening items more consistent. Fixes #544
The affect on selection when using File > Open vs. double-clicking
differed, for both volumes and folders. For the menu item (or via the
shortcut key) the selection would be cleared, whereas for
double-clicking it wouldn't be. This was added to work around the
weirdness of having selection in an inactive window, but was not done
consistently.

Update the code to a more consistent behavior:

* If selection is on the desktop (volume icons), then leave selection
  intact. This is because desktop icons can always be repainted
  correctly.

* If selection is in window (folder icons), then clear selection. This
  is because content in inactive windows can't be arbitrarily
  repainted.