DeskTop: Clear file selection immediately on File > Open. Fixes #415
Logic was added to clear selection when a folder was opened, but this
was done after the window appeared. If the window obscured the icon
this was fine, since the icon would repaint correctly when the window
was moved. But if the window did not obscure the icon, it would still
appear selected.

Clear the selection at the start of processing the command if needed
(only applies to folders, not volumes or other files) and use a copy
of the selected icon list to complete the operation.