DeskTop: Tweak icon updating during rename. Prep work for not #469.
When renaming an icon, DeskTop would attempt to erase the icon and
update the name. This was broken in a couple of ways. IconTK's
EraseIcon didn't calculate the icon poly, so it would only if the poly
was correct, e.g. for the last selected volume icon. And DeskTop
wouldn't redraw the icon.

Fix EraseIcon to calc the poly, and add code to redraw the icon.
Before the name is updated, re-case it.

Note that this is all fairly moot, because after the rename the
containing window is repainted multiple times and then reloaded, and
renamed volume icons are destroyed and recreated. In a fast emulator
you wouldn't notice. At 1MHz the redundant repaints are pretty sad.