DeskTop: Fix drag selection edge case. Fixes #551
Logic was added in cba366d7 to allow drag-select in a window with
modifier to not clear selection. But if selection was on the desktop
(volume icons), this would leave selection in a bad state. Add a test
to only skip clearing if selection is already in the active window.