DeskTop: Add File > Duplicate... command. Fixes #228
This shows a dialog rather than autogenerating a name like FILE2 as
that was easier. The dialog is basically the same as Rename or New
Folder, and prompts for a name.

The Apple-D shortcut is provided, but there's no drag gesture to
auto-duplicate because you need to enter a name anyway.