DeskTop: Fix un-shading of parent icons when closing windows
When closing windows, the old logic was:

* Close window (leaving a blank spot on screen)
* Find the "parent icon" (folder/volume) for the window
* Mark it not open (not shaded)
* Mark it selected (highlighted)
* Paint the icon (1)
* Animate the window closing
* Clear updates (fill in the blank spot)
* Force a full window repaint if selection changed (2)

This would fail in various ways. Painting the icon (1) would generally
fail for windowed icons, and could be overdrawn (fully or partially)
when updates were cleared. Forcing a full repaint would never happen
since selection couldn't be in a background. Rework this to be:

* Close window (leaving a blank spot on screen)
* Find the "parent icon" (folder/volume) for the window
* Animate the window closing
* Clear updates (fill in the blank spot)
* Mark it not open (not shaded)
* Mark it selected (highlighted)
* Paint the icon

This handles folder icons in the top window, and volume icons. Icons
in background windows will show incorrectly until they get repainted
somehow.