DAs: Manually clear update events on close
DeskTop is changed to not force a "clear updates" pass when the DA is
closed. This improves performance for DAs that didn't close windows,
since otherwise all the volume icons would need to be repainted. It
makes for slightly more book-keeping for DAs, but they need to use the
toolkit correctly anyway.