DeskTop: Handle MGTK window_obscured error. Fixes #555
Any time DeskTop asks for a window port (GetWinPort) it's possible
that the result is an error, because the window is offscreen. This
needs to be handled in every path.

Change the procs that call GetWinPort to pass along the error, and all
callers to skip painting in such cases. This required a fair amount of
work because the IconTK calls that do book-keeping (AddIcon,
RemoveIcon, HighlightIcon, UnhighlightIcon) changed state and also
painted. So those APIs were changed to not paint, and the calling
sites were untangled.

The main methods were given an `unsafe` prefix so callers know to
check. A `safe` version is available for call sites to assert that the
window will never be obscured, e.g. a non-movable dialog.

Call sites touched:
* Processing redraws when an overlapping window closes
* Selecting the vol icon when a window is activated
* Switching a window to list view
* Switching a window to icon view
* Highlighting icons with the keyboard
* Select All
* Scrolling by moving the elevator
* Selecting file icons by name (e.g. new folder, window close)
* Selecting/deselecting file icons by clicking
* Refreshing a window after content change (e.g. adding/deleting a file)
* Updating the parent icon when a window is closed
* Scrolling by clicking the arrows/pagers
* Opening a window from an icon
* Marking a volume/folder icon as open
* Opening a window from a path
* Drawing the entries for a window
* Clearing selection
* Marking a volume/folder icon as not open
* Renaming an icon

As you can see, there are lot of seemingly redundant places that draw
windows or icons. There's plenty of potential for follow-up work to
dedupe code. While not every possible cleanup was done in this change,
various places that did extraneous painting were optimized (e.g.
window scrolling no longer redraws the window header), some redundant
methods were merged, and common routines were factored out. This saves
~380 bytes in main.