Add a new DA "Darkness"
This DA sits in the "Optional" directory, and it's really best used
by adding it to the Selector menu so it can be launched with a keyboard
shortcut (e.g. Apple-8).

The DA paints over the entire screen with a dark pattern then exits.

This is a useful debugging tool when looking for extraneous repaints.
For example, open a window, put the cursor over an icon, invoke the
DA, then click an icon in the window. Did only the icon repaint, or
did extra work happen? Put the cursor over the title bar of the
window, invoke the DA, then drag the window. Did extra work happen?

Because this is really only of use for testers and developers,
it will not appear in the menu for most users.