IconTK: Use a state bit to track highlighting.
Various operations (e.g. RedrawIcon) need to check to see if an icon
is highlighted. This was done with a search through a list. While the
list still needs to be maintained (e.g. to drag all highlighted), use
a bit in the state instead to save searching in many cases.

Note that a "is this icon in the allocated list?" tests still exist
and are checked for most (!) calls. Potentially replace that with a
table for a perf boost.

Also, dedupe some logic and correct bad disasm about a state bit being
used while dragging.