DeskTop: Simplify setting used/free for windows opened by path
The used/free logic is a bit convoluted; DeskTop originally queried
used/free from a volume when the volume icon was opened. Child windows
could re-use what the parent already had to avoid querying the volume.
That's nice, but doesn't work if the window is opened by path, e.g. on
restore. So when open-by-path logic was added, the window would end up
with null values and then be overpainted with bespoke code.

Change DeskTop to always query when opening a window, which is a tiny
speed hit but allows the open-by-path to be much simpler.