DeskTop: Patch by @apple2geek to fix format for large volumes. Fixes #557
The logic for constructing the volume bitmap was flawed. It assumed
that the block size of the volume would be an even multiple of 8, and
that the maximum volume bitmap size would be 10 blocks. If these
assumptons were violated the bitmap was wrong. This would result in
behavior such as the first file written to disk overlapping the
bitmap, causing portions of the disk to become unavailable, and the
file becoming corrupted.

The new code (c/o @apple2geek) handles these cases. It was tested with
volumes of the following sizes (in blocks), using cadius CHECKVOLUME:

280
1600
4095
4096
4097
16383
16384
16385
49151
49152
49153
65534
65535