DeskTop: Improve driver address handling for Format/Erase
The Format/Erase logic had a convoluted way to get the driver address.
Replace this with a much simpler form, and put it in a proc.

Also, the code made assumptions about the driver being firmware
($CnXX) if the unit number wasn't $X0, which is not valid e.g. for
remapped drives. Make the logic more robust.