DeskTop: Clear "move" flag when doing explicit copies. Fixes #562
When moving a file (dragging to same volume, etc) a flag is set at the
start of the operation, then checked at the end of the operation and
the source is deleted if set. Unfortunately, that same check at the
end of the operation is run for two copy cases: the explicit File >
Copy a File... operation, and the the implicit "Copying to RAMCard..."
logic for Selector items. This caused the original to be deleted.

Ensure the flag is cleared at the start of those copy operations.