DAs: Add Electric Duet player
This plays files from Electric Duet by Paul Lutus, using a player
coded by Alexander Patalenski. The files must be converted to ProDOS
and given type $D5 and auxtype $D0E7 to show up with the right
icon and be previewable.

If you have files without that type set, you can copy the DA
to the DESK.ACC folder and invoke it manually.