DeskTop: Only show "save the new list..." prompt if needed. Fixes #567
When DeskTop is copied to RAMCard, and you modify Selector entries,
DeskTop writes the updated SELECTOR.LIST to the copy on the ramdisk,
but then prompts "Do you want to save the new list on the system
disk?" and only if you select OK does it try to write to the original
fdir. This is original MouseDesk behavior, and makes sense if you are
likely to have swapped out your system disk. If you have swapped out
the original disk you'll get a retry prompt.

In a fixed-disk scenario where you are unlikely to have removed the
original and running from RAM is just a speed optimization, prompting
is annoying. But on the other hand, if you have removed the disk,
having the changes be silently lost is also annoying.

Change the logic to silently try updating the file with the original
path, and only show the prompt if that fails. If the user inserts the
system disk before clicking OK, no more prompts will show. Otherwise a
"Please insert the system disk" prompt will show.