Date DA: Prompt to save on the system disk if needed.
Follow on to 7efd38b8 and 8f8a1a3f for #567 - have the Date DA also
try to write the date to the system disk if running from a RAMCard.
Previously it would only write to the copy on the RAMCard.

This only applies if the system doesn't have a real-time clock/driver.