MGTK/DeskTop: Introduce update event for desktop itself. #554
When erasing a window (e.g. in response to a move/close/resize), MGTK
will now emit an "update" event for window id 0 i.e. the desktop
itself.

This allows clients (like, say, DeskTop) to paint onto the desktop
before the first window is drawn. The usual BeginUpdate/ EndUpdate
calls can be made passing in window id 0. This will result in a
grafport clipped to the update region, so that minimal repaints can be
done.