Selector: Fix crash from alert dialog. Fixes #575
The alert code called back into the app to yield during the event
loop, but the alert code is in Aux LCBANK1 so that was left paged in.
When clock support was added to the yield, this caused a crash since
ProDOS GET_TIME was invoked.

Add a relay that banks the main ZP/ROM in while yielding.