Disk Copy: Don't clear keyboard strobe between each pass. Fixes #429
The Disk Copy routine watches for an Esc keypress to cancel the copy.
But it clears the keyboard strobe at the start of each set of blocks,
so if you hit Esc in a window between the check and a set of blocks,
the keypress is ignored.

There's no need to clear the strobe, so just don't. It's cleared after
a keypress during the pass, which seems to be all that's needed.