MGTK: Fix drawing title bars with no room between close box and name
If there was no room to draw the lines between the close box and
title in a window header, MGTK would end up with a bad rect and
trash the screen. Add in a check.

This allows using the "Calcolatrice" title for the Calc in IT.
Also, pick up some other L10N changes.