DeskTop: Name renaming a file with the same name a no-op. Fixes #584
Following ebaf58cf the Rename dialog would show an alert if the name
is already in use. This prevented a "no-op" rename, just hitting OK.
Since that's convenient, especially in a bulk rename, re-enable this
with explicit code (the rename is not actually attempted, no
icon/window updates are made, etc).

As a side effect, this fixes #585 as well, which was caused by trying
to update window paths when nothing had changed.