DeskTop: Rename "Selector" menu to "Shortcuts". Fixes #378
This updates the UI strings referencing the "Selector" menu to be
"Shortcuts", which I believe does a much better job of explaining what
it's good for. And "shortcut" is much clearer and shorter than
"selector entry".

The file is still called SELECTOR.LIST (for now), and the optional
startup application is still called "Selector" which seems fine.