DeskTop/Disk Copy: Make Disk II identification more robust.
Both DeskTop and Disk Copy used an obsolte heuristic that the low byte
of a unit number signified a device type, with $x0 meaning Disk II.
Per ProDOS 8 Technical Note #21, that's invalid; the low nibble is
just derived from the $CnFE firmware byte from the device, or
something arbitrary.

Further, I'd misinterpreted the meaning of part of TN#21 which said
to use the SSS bytes of the unit number. I thought that was bogus
for non-firmware drivers, but it's valid for Disk II devices which have
non-firmware drivers! If there's a firmware driver, and the SSS bits
point at a ProDOS device that's a Disk II, it *is* valid since Disk II
devices are never remapped.

Note that there are still fiddly cases; if the SSS bits don't point
at a ProDOS device, we can't further identify it. If the SSS bits do
point at a ProDOS device, then in theory it shouldn't be remapped.

Anyway, this means we can construct a rigorous "does this unit number
refer to a Disk II?" method, and use it in DeskTop's icon creation,
the Format/Erase device naming and format logic, and Disk Copy's
device-specific logic, removing reliance on a "device type nibble"

Reference:
http://www.1000bit.it/support/manuali/apple/technotes/pdos/tn.pdos.21.html