DeskTop: Make alert messages wrap automatically, rather than manually.
Previously, "warnings" (a special case of alerts for DeskTop)
supported longer two-line messages. These needed to be manually
wrapped, potentially confusing localizers e.g. for Spanish where ¿ and
? would appear in different strings.

Teach the alert code in DeskTop (and the copy in Selector) to wrap the
string automatically, although it only handles a single break.