L10n: Localize desk accessory names in disk images
* Add translation resource entries for the DA filenames

* When creating disk images with `make package`, make the DA
  folders/names using the table

The names must follow ProDOS conventions (so A-Z0-9. only) and the
translations have not yet been reviewed by localizers yet, so they're
going to be really bad. It takes a native speaker to figure out how to
abbreviate appropriately.

Also note that this does not yet apply to other make targets (install,
mount, shk, ...).