Disk Copy: Improve Disk Copy ejectability test
Previous test used the low nibble of the unit number as a "device
type". Switch to probing via SmartPort call to see if it's a 3.5"
disk.

This also unifies the logic for determining SmartPort unit numbers
between DeskTop and Disk Copy, replacing a previous ProDOS 1.x-only
mechanism with something that's ProDOS 2.x-friendly.