DeskTop: Fix corruption after switching to list views. Fixes #592
The "nominal" state for DeskTop is that the "cached" icon table is for
the desktop. Various operations will load in the icon table for
windows, and are required to save (if modified) and load in the
desktop's icon table when complete. This was missing for the "change
to list view" operation, so the wrong table would be around.

Changing selection using the arrow keys uses the active window's icon
table (if in icon view) and the desktop's icon table to construct a
list of plausible icons. If the active window was in list view, only
the desktop's table was used... but if the view had just changed it
wouldn't be loaded, leading to a bogus list, and sadness.

Add the missing call at the end of the list view operation, and audit
all other call sites to ensure any window table load is finished off
by a desktop table load.