Launcher: Fix minor free-space calculation bug noticed via inspection.
When copying to RAMCard during launch, an attempt is made to check for
free space before each file is copied. This is done by querying the
volume for free space: free = total - used. And then ensure that the
new_size is < free. To account for overwriting, where the new file is
going to replace an old file with the same name, this should really
be: free = (total - used) + old_size. The logic was inverted, though.

This is moot (and the bug was probably never noticed) because the
launcher doesn't handle overwriting! If the destination file is
present, the copy will silently fail. But fix it anyway.